social_scraper/requirements.txt
requests
elasticsearch
schedule
python-dotenv
